requirements_tensorflow.txt
absl-py==1.4.0
array-record==0.5.0
astunparse==1.6.3
beartype==0.17.2
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
clu==0.0.11
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
dm-tree==0.1.8
equinox==0.11.3
etils==1.7.0
flatbuffers==24.3.7
flax==0.8.2
fonttools==4.50.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
gviz-api==1.10.0
h5py==3.10.0
idna==3.6
importlib_resources==6.4.0
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
jax==0.4.25
jaxdf==0.2.7
jaxlib==0.4.25+cuda11.cudnn86
jaxtyping==0.2.28
jwave==0.2.0
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.3.2
msgpack==1.0.8
namex==0.0.7
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.9.6.50
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
opt-einsum==3.3.0
optax==0.2.1
optree==0.10.0
orbax-checkpoint==0.5.7
packaging==24.0
pillow==10.2.0
plum-dispatch==2.3.2
promise==2.3
protobuf==3.20.3
psutil==5.9.8
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scipy==1.12.0
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard_plugin_profile==2.15.1
tensorflow==2.16.1
tensorflow-datasets==4.9.4
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-metadata==1.14.0
tensorstore==0.1.56
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tqdm==4.66.2
typeguard==2.13.3
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1
